data-engineer-challenge/iot_data_generator/requirements.txt
annotated-types==0.7.0
dnspython==2.7.0
paho-mqtt==2.1.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.11.1
python-dotenv==1.0.1
typing_extensions==4.12.2